Fix claim preview and let it ready to use in another place
