+++ src/pagePartials/badge/preview/BadgeView.tsx
@@ -1,0 +1,41 @@
+import * as React from 'react'
+
+import { BadgePreview } from '@thebadge/ui-library'
+
+import useBadgeModel from '@/src/hooks/subgraph/useBadgeModel'
+import { getBackgroundBadgeUrl } from '@/src/utils/badges/getBackgroundBadgeUrl'
+import { BadgeNFTAttributesType } from '@/types/badges/BadgeMetadata'
+
+type BadgePreviewGeneratorProps = {
+  modelId: string
+  badgeUrl?: string
+  additionalData?: Record<string, any>
+}
+
+export const BadgeView = ({ badgeUrl, modelId }: BadgePreviewGeneratorProps) => {
+  const badgeModelData = useBadgeModel(modelId)
+  const badgeModelMetadata = badgeModelData.data?.badgeModelMetadata
+  const badgeLogoImage = badgeModelData.data?.badgeModelMetadata?.image
+
+  const backgroundType = badgeModelMetadata?.attributes?.find(
+    (at) => at.trait_type === BadgeNFTAttributesType.Background,
+  )
+
+  const textContrast = badgeModelMetadata?.attributes?.find(
+    (at) => at.trait_type === BadgeNFTAttributesType.TextContrast,
+  )
+
+  return (
+    <BadgePreview
+      animationEffects={['wobble', 'grow', 'glare']}
+      animationOnHover
+      badgeBackgroundUrl={getBackgroundBadgeUrl(backgroundType?.value)}
+      badgeUrl={badgeUrl}
+      category={badgeModelMetadata?.name}
+      description={badgeModelMetadata?.description}
+      imageUrl={badgeLogoImage?.s3Url}
+      size="medium"
+      textContrast={textContrast?.value || 'light-withTextBackground'}
+    />
+  )
+}
