+++ src/pagePartials/badge/preview/DiplomaView.tsx
@@ -1,13 +1,10 @@
-import React, { useCallback, useEffect, useRef } from 'react'
+import React from 'react'
 
-import { Box } from '@mui/material'
 import { DiplomaPreview } from '@thebadge/ui-library'
-import { UseFormSetValue } from 'react-hook-form'
 
 import { APP_URL } from '@/src/constants/common'
 import useBadgeModel from '@/src/hooks/subgraph/useBadgeModel'
 import useS3Metadata from '@/src/hooks/useS3Metadata'
-import { convertPreviewToImage } from '@/src/pagePartials/badge/mint/utils'
 import enrichTextWithValues, { EnrichTextValues } from '@/src/utils/enrichTextWithValues'
 import {
   DiplomaFooterConfig,
@@ -19,13 +16,11 @@ import { BackendFileResponse } from '@/types/utils'
 
 type Props = {
   modelId: string
-  setValue: UseFormSetValue<any>
   badgeUrl?: string
   additionalData?: Record<string, any>
 }
 
-export default function DiplomaPreviewGenerator({ additionalData, modelId, setValue }: Props) {
-  const diplomaPreviewRef = useRef<HTMLDivElement>()
+export default function DiplomaView({ additionalData, modelId }: Props) {
   const badgeModelData = useBadgeModel(modelId)
   const badgeModelMetadata = badgeModelData.data?.badgeModelMetadata
 
@@ -59,20 +54,6 @@ export default function DiplomaPreviewGenerator({ additionalData, modelId, setVa
     content: DiplomaIssuerConfig<BackendFileResponse>
   }>((issuerConfigs?.value as string) || '')
 
-  const generatePreviewImage = useCallback(
-    async (badgePreviewRef: React.MutableRefObject<HTMLDivElement | undefined>) => {
-      const previewImage = await convertPreviewToImage(badgePreviewRef)
-      setValue('previewImage', previewImage)
-    },
-    [setValue],
-  )
-
-  useEffect(() => {
-    if (diplomaPreviewRef.current) {
-      generatePreviewImage(diplomaPreviewRef)
-    }
-  }, [diplomaPreviewRef, generatePreviewImage])
-
   const issuerAvatarUrl = issuerConfigsMetadata.data?.content.issuerAvatar?.s3Url
   const issuerLabel = issuerConfigsMetadata.data?.content.issuedByLabel
   const issuerTitle = issuerConfigsMetadata.data?.content.issuerTitle
@@ -91,32 +72,26 @@ export default function DiplomaPreviewGenerator({ additionalData, modelId, setVa
   const footerText = footerConfigsMetadata.data?.content.footerText
 
   return (
-    <>
-      <Box ref={diplomaPreviewRef}>
-        <DiplomaPreview
-          animationEffects={['wobble', 'grow', 'glare']}
-          animationOnHover
-          backgroundUrl={'https://dev-app.thebadge.xyz/shareable/diploma-background.png'}
-          badgeUrl={APP_URL}
-          courseName={courseName?.value}
-          date={achievementDate?.value}
-          decorationBackgroundUrl={'https://dev-app.thebadge.xyz/shareable/diploma-decoration.png'}
-          description={achievementDescription?.value}
-          footerText={
-            footerEnabled
-              ? enrichTextWithValues(footerText, additionalData as EnrichTextValues)
-              : ''
-          }
-          issuedByLabel={issuerLabel || 'Issued by'}
-          issuerAvatarUrl={issuerAvatarUrl}
-          issuerDescription={issuerDescription}
-          issuerIsVerified={''}
-          issuerTitle={issuerTitle}
-          studentName={enrichTextWithValues('{{studentName}}', additionalData as EnrichTextValues)}
-          textContrastRight="dark"
-          {...signatureProps}
-        />
-      </Box>
-    </>
+    <DiplomaPreview
+      animationEffects={['wobble', 'grow', 'glare']}
+      animationOnHover
+      backgroundUrl={'https://dev-app.thebadge.xyz/shareable/diploma-background.png'}
+      badgeUrl={APP_URL}
+      courseName={courseName?.value}
+      date={achievementDate?.value}
+      decorationBackgroundUrl={'https://dev-app.thebadge.xyz/shareable/diploma-decoration.png'}
+      description={achievementDescription?.value}
+      footerText={
+        footerEnabled ? enrichTextWithValues(footerText, additionalData as EnrichTextValues) : ''
+      }
+      issuedByLabel={issuerLabel || 'Issued by'}
+      issuerAvatarUrl={issuerAvatarUrl}
+      issuerDescription={issuerDescription}
+      issuerIsVerified={''}
+      issuerTitle={issuerTitle}
+      studentName={enrichTextWithValues('{{studentName}}', additionalData as EnrichTextValues)}
+      textContrastRight="dark"
+      {...signatureProps}
+    />
   )
 }
